Quick_type_bug_fix
Fixed a small bug where num_slices was getting stored as a float instead of an int.

+++ bxh2bids.py
@@ -247,7 +247,7 @@ def create_bold_json(bxh_file, full_output):
             else:
                 logging.warn('Cannot find slice order in bxh file!')
                 logging.warn('Assuming it is sequential!')
-                num_slices = float(element['size'])
+                num_slices = int(element['size'])
                 slice_order_list = range(1, num_slices+1)
     factor = tr/num_slices
     st = []
